washedmcp/pyproject.toml
[project]
name = "cleanmcp"
version = "1.0.0"
description = "MCP Recommender and Auto-Installer with JSON-based logging"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "google-generativeai>=0.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tools", "server.py", "json_storage.py"]
